Reduce some duplicity
