+++ src/main/sagas/servers.js
@@ -1,44 +1,9 @@
-import fs from 'fs';
-import path from 'path';
-
-import { app } from 'electron';
 import { call, delay, put, race, select } from 'redux-saga/effects';
 
 import { SERVERS_READY } from '../../actions';
 import { readFromStorage, readItem, removeItem } from '../localStorage';
 import { selectServers, selectCurrentServerUrl } from '../selectors';
-
-const getConfigurationPath = (filePath, { appData = true } = {}) => path.join(
-	...appData ? [
-		app.getAppPath(),
-		app.getAppPath().endsWith('app.asar') ? '..' : '.',
-	] : [app.getPath('userData')],
-	filePath,
-);
-
-const readConfigurationFile = async (filePath, {
-	appData = true,
-	purgeAfter = false,
-} = {}) => {
-	try {
-		const configurationFilePath = getConfigurationPath(filePath, { appData });
-
-		if (!await fs.promises.stat(filePath).then((stat) => stat.isFile(), () => false)) {
-			return null;
-		}
-
-		const content = JSON.parse(await fs.promises.readFile(configurationFilePath, 'utf8'));
-
-		if (!appData && purgeAfter) {
-			await fs.promises.unlink(configurationFilePath);
-		}
-
-		return content;
-	} catch (error) {
-		console.warn(error);
-		return null;
-	}
-};
+import { readConfigurationFile } from '../fileSystemStorage';
 
 export function *validateServerUrl(serverUrl, timeout = 5000) {
 	const url = new URL(serverUrl);
