+++ src/main/sagas/updates.js
@@ -1,6 +1,3 @@
-import fs from 'fs';
-import path from 'path';
-
 import { app } from 'electron';
 import { autoUpdater, CancellationToken } from 'electron-updater';
 import { select, call, put, takeEvery } from 'redux-saga/effects';
@@ -23,38 +20,7 @@ import {
 	selectSkippedUpdateVersion,
 } from '../selectors';
 import { readFromStorage } from '../localStorage';
-
-const getConfigurationPath = (filePath, { appData = true } = {}) => path.join(
-	...appData ? [
-		app.getAppPath(),
-		app.getAppPath().endsWith('app.asar') ? '..' : '.',
-	] : [app.getPath('userData')],
-	filePath,
-);
-
-const readConfigurationFile = async (filePath, {
-	appData = true,
-	purgeAfter = false,
-} = {}) => {
-	try {
-		const configurationFilePath = getConfigurationPath(filePath, { appData });
-
-		if (!await fs.promises.stat(filePath).then((stat) => stat.isFile(), () => false)) {
-			return null;
-		}
-
-		const content = JSON.parse(await fs.promises.readFile(configurationFilePath, 'utf8'));
-
-		if (!appData && purgeAfter) {
-			await fs.promises.unlink(configurationFilePath);
-		}
-
-		return content;
-	} catch (error) {
-		console.warn(error);
-		return null;
-	}
-};
+import { readConfigurationFile } from '../fileSystemStorage';
 
 const isUpdatingAllowed = (process.platform === 'linux' && !!process.env.APPIMAGE)
 	|| (process.platform === 'win32' && !process.windowsStore)
